requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
numpy==1.26.4
pillow==10.3.0
opencv-python-headless==4.9.0.80
mobile-sam==0.1.1
torch==2.3.0
torchvision==0.18.0
